rust: SequenceNumber struct
